Move type defs from templates.js to types.flow.js
